requirements.txt
alabaster==0.7.13
aniso8601==9.0.1
apipkg==3.0.1
apparmor==3.1.2
appdirs==1.4.4
application-utility==1.3.2
apsw==3.40.0.0
asgiref==3.6.0
astroid==2.14.2
async-generator==1.10
async-timeout==4.0.2
attrs==22.2.0
autocommand==2.2.2
Babel==2.11.0
bcrypt==4.0.1
beautifulsoup4==4.11.1
Brotli==1.0.9
brotlicffi==1.0.9.2
bs4==0.0.1
btrfsutil==6.1.3
cachetools==5.3.0
catfish==4.16.4
cchardet==2.1.7
certifi==2022.12.7
cffi==1.15.1
chardet==5.1.0
cheroot==9.0.0
CherryPy==18.8.0
click==8.1.3
colorama==0.4.6
contourpy==1.0.7
coverage==6.5.0
cryptography==39.0.1
css-parser==1.0.8
cssselect==1.2.0
cycler==0.11.0
Cython==0.29.33
defusedxml==0.7.1
dill==0.3.6
distlib==0.3.6
distro==1.8.0
Django==4.1.7
django-ckeditor==6.5.1
django-embed-video==1.4.8
django-js-asset==2.0.0
django-rename-app==0.1.5
django-taggit==3.1.0
dnspython==2.2.1
docopt==0.6.2
docutils==0.19
easy-thumbnails==2.8.5
elasticsearch==7.9.0
exceptiongroup==1.1.0
fastjsonschema==2.16.2
feedparser==6.0.10
filelock==3.8.2
Flask==2.2.2
Flask-RESTful==0.3.9
fonttools==4.38.0
gitdb==4.0.10
GitPython==3.1.29
google-api-core==2.11.0
google-api-python-client==2.70.0
google-auth==2.16.0
google-auth-httplib2==0.1.0
googleapis-common-protos==1.58.0
grpcio==1.51.1
gufw==22.4.0
h11==0.14.0
html2text==2020.1.16
html5-parser==0.4.10
html5lib==1.1
httplib2==0.21.0
idna==3.4
ifaddr==0.2.0
imagesize==1.4.1
importlib-metadata==5.0.0
inflate64==0.3.1
inflect==6.0.2
iniconfig==2.0.0
isodate==0.6.1
isort==5.12.0
itsdangerous==2.1.2
jaraco.classes==3.2.3
jaraco.collections==3.5.2
jaraco.context==4.3.0
jaraco.functools==3.5.2
jaraco.text==3.11.1
jeepney==0.8.0
Jinja2==3.1.2
joblib==1.2.0
keyutils==0.6
kiwisolver==1.4.4
lazy-object-proxy==1.9.0
lensfun==0.3.3
LibAppArmor==3.1.2
libtorrent==2.0.8
lightdm-gtk-greeter-settings==1.2.2
lit==15.0.7.dev0
logutils==0.3.5
louis==3.24.0
lxml==4.9.2
Mako==1.2.4
manjaro-sdk==0.1
Markdown==3.4.1
MarkupSafe==2.1.2
matplotlib==3.6.2
mccabe==0.7.0
mechanize==0.4.8
menulibre==2.2.3
mock==3.0.5
more-itertools==9.0.0
msgpack==1.0.4
mugshot==0.4.3
multivolumefile==0.2.3
mutagen==1.46.0
netifaces==0.11.0
netsnmp-python==1.0a1
nftables==0.1
norminette==3.3.51
nose==1.3.7
npyscreen==4.10.5
nspektr==0.4.0
numpy==1.24.1
ordered-set==4.1.0
outcome==1.2.0
packaging==23.0
pacman-mirrors==4.23.2
paperclip==2.6.1
pbr==5.11.1
pdftotext==2.2.2
pecan==1.4.2
pexpect==4.8.0
Pillow==9.4.0
Pivy==0.6.8
pkgconfig==1.5.5
platformdirs==2.6.2
pluggy==1.0.0
ply==3.11
pooch==1.6.0
portend==3.1.0
prettytable==3.6.0
prometheus-client==0.15.0
protobuf==4.21.12
psutil==5.9.4
ptyprocess==0.7.0
py==1.11.1.dev0+g447bac51.d20230131
py7zr==0.20.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pybcj==1.0.1
pycairo==1.23.0
pychm==0.8.6
pycodestyle==2.10.0
pycparser==2.21
pycryptodome==3.17
pycryptodomex==3.12.0
pydantic==1.10.4
Pygments==2.14.0
PyGObject==3.42.2
PyInputPlus==0.2.12
PyJWT==2.6.0
pylint==2.16.2
pyOpenSSL==23.0.0
pyparsing==3.0.9
pyperclip==1.8.2
pyppmd==1.0.0
PyQt5==5.15.9
PyQt5-sip==12.11.1
PyQt6==6.4.2
PyQt6-sip==13.4.1
PyQt6-WebEngine==6.4.0
PySide2==5.15.8
PySide6==6.4.2
PySimpleValidate==0.2.12
pysmbc==1.0.23
PySocks==1.7.1
pytest==7.2.1
pytest-cov==4.0.0
python-dateutil==2.8.2
python-decouple==3.7
pytz==2022.7
pyxdg==0.28
PyYAML==6.0
pyzstd==0.15.3
regex==2022.10.31
reportlab==3.6.12
repoze.lru==0.7
requests==2.28.1
Routes==2.5.1
rsa==4.9
scikit-learn==1.2.0
scipy==1.10.0
selenium==4.8.2
Send2Trash==1.8.0
sgmllib3k==1.0.0
shiboken2==5.15.8
shiboken6==6.4.2
shiboken6-generator==6.4.2
shtab==1.5.8
six==1.16.0
smmap==5.0.0
sniffio==1.3.0
snowballstemmer==2.2.0
sortedcontainers==2.4.0
soupsieve==2.3.2.post1
speedtest-cli==2.1.3
Sphinx==5.3.0
sphinxcontrib-applehelp==1.0.4
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==2.0.1
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.5
sqlparse==0.4.3
stdiomask==0.0.6
systemd-python==235
TBB==0.2
team==1.0
tempora==5.2.0
termcolor==2.2.0
texttable==1.6.7
threadpoolctl==3.1.0
tldr==3.1.0
toml==0.10.2
tomli==2.0.1
tomlkit==0.11.6
trio==0.22.0
trio-websocket==0.9.2
trove-classifiers==2023.2.9
typing_extensions==4.4.0
udiskie==2.4.2
ufw==0.36.1
unrardll==0.1.5
uritemplate==4.1.1
urllib3==1.26.12
validate-pyproject==0.12.1
virtualenv==20.17.1
waitress==2.1.2
wcwidth==0.2.5
webencodings==0.5.1
WebOb==1.8.7
websockets==10.4
WebTest==3.0.0
Werkzeug==2.2.2
wrapt==1.14.1
wsproto==1.2.0
xmlsec==1.3.13
yt-dlp==2023.3.4
zc.lockfile==2.0
zeroconf==0.39.4
zipp==3.13.0
